fix: unify user_input field naming to client_type/client_version
- docs/api.md: fix /api/assistant request body fields (platform→client_type, app_version→client_version)
- web-admin: rename UserInput fields in FunctionTester, DagEditor, function.ts, workflow.ts
- game_service: fix INT UNSIGNED column decoded as i64 → u32 in list_external_games
- generate_answer_node: fix experimental guard pattern to stable match arm guard

+++ crates/hiveweb/src/runtime/builtins/game_list.rs
@@ -46,7 +46,7 @@ async fn game_list_async_impl(
         .map_err(|e| BuiltinError::Exec(format!("{e}")))?;
 
     // 3. Merge: for each external game, collect all aliases (its own + internal supplements)
-    let mut entries: Vec<(i64, String, Vec<String>)> = Vec::new();
+    let mut entries: Vec<(u32, String, Vec<String>)> = Vec::new();
     for (id, name, ext_alias) in external {
         let mut aliases: Vec<String> = Vec::new();
